# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reminders:
    get:
      operationId: getReminders
      tags: [reminders]
      summary: List all reminders
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderBody"
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    get:
      operationId: getReminder
      tags: [reminders]
      summary: Get a reminder by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
        "404":
          description: Not found
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReminderBody"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /voice-agent/config:
    get:
      operationId: getVoiceAgentConfig
      tags: [voice-agent]
      summary: Get voice agent configuration
      responses:
        "200":
          description: Voice agent configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceAgentConfig"
    patch:
      operationId: updateVoiceAgentConfig
      tags: [voice-agent]
      summary: Update voice agent configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVoiceAgentConfigBody"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceAgentConfig"
  /vapi/webhook:
    post:
      operationId: handleVapiWebhook
      tags: [voice-agent]
      summary: Receive Vapi call event webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                required:
                  - received
components:
  schemas:
...
    Reminder:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        type:
          type: string
          enum: [medication, visit, appointment, follow_up, wellness_check]
        message:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [pending, sent, acknowledged, failed, cancelled]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientId
        - type
        - message
        - scheduledAt
        - status
        - createdAt
        - updatedAt
    CreateReminderBody:
      type: object
      properties:
        clientId:
          type: integer
        type:
          type: string
          enum: [medication, visit, appointment, follow_up, wellness_check]
        message:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [pending, sent, acknowledged, failed, cancelled]
      required:
        - clientId
        - type
        - message
        - scheduledAt
    UpdateReminderBody:
      type: object
      properties:
        type:
          type: string
          enum: [medication, visit, appointment, follow_up, wellness_check]
        message:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [pending, sent, acknowledged, failed, cancelled]
...
    VoiceAgentConfig:
      type: object
      properties:
        id:
          type: integer
        assistantName:
          type: string
        systemPrompt:
          type: string
        supportedLanguages:
          type: string
        escalationPhone:
          type: string
        vapiAssistantId:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxCallDurationSeconds:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - assistantName
        - systemPrompt
        - supportedLanguages
        - escalationPhone
        - isActive
        - maxCallDurationSeconds
        - createdAt
        - updatedAt
    UpdateVoiceAgentConfigBody:
      type: object
      properties:
        assistantName:
          type: string
        systemPrompt:
          type: string
        supportedLanguages:
          type: string
        escalationPhone:
          type: string
        vapiAssistantId:
          type: ["string", "null"]
        isActive:
          type: boolean
        maxCallDurationSeconds:
          type: integer